requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
langsmith>=0.2.0
fastapi>=0.115.0
uvicorn>=0.34.0
requests>=2.32.0
langgraph-checkpoint-sqlite>=2.0.0
